ref(onboarding): Move onboardingtask mutation hook to live in utils
